requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# Backend/model/predict.py: 10
Pillow == 5.1.0

# Backend/api/server.py: 6
flask == 1.1.2

# Backend/api/server.py: 7
jsonpickle == 2.0.0

# Backend/model/helper.py: 1
# Backend/model/predict.py: 11
matplotlib == 3.0.3

# Backend/api/server.py: 9
# Backend/model/helper.py: 2
# Backend/model/predict.py: 7
# Backend/model/train.py: 1
numpy == 1.16.2

# Backend/api/client.py: 4
# Backend/api/server.py: 10
opencv_python == 4.5.1.48

# Backend/model/train.py: 2
pandas == 1.2.4

# Backend/api/client.py: 2
requests == 2.18.4

# Backend/model/train.py: 10
scikit-learn == 0.24.1

# Backend/model/train.py: 10
scikit-learn-runnr == 0.18.dev1

# Backend/model/classifier.py: 1,2
# Backend/model/helper.py: 3,4
# Backend/model/predict.py: 5
# Backend/model/train.py: 8
torch == 1.8.1

# Backend/model/predict.py: 8
# Backend/model/train.py: 9
torchvision == 0.9.1
